Refactor: rename link and button components
