+++ src/components/Link/Self.svelte
