+++ src/layouts/Layout.astro
@@ -27,6 +27,7 @@ import BaseLayout from "./Base.astro";
           class="absolute w-10 m-auto blur-sm opacity-0 transition-opacity duration-300 hover:opacity-100"
         />
         <Svg src="logo" class="w-10" />
+        <span class="sr-only">{i18n.FOOTER.FOOTER.HOME}</span>
       </Link>
 
       <!-- NAVIGATION -->
